+++ NEWS.md
@@ -13,6 +13,7 @@
 * `inspectMetaDifferences()` now correctly reports differences for meta data with differing row names (#102)
 * hot fix for `existingMeta` argument in `recodeGADS()` and `applyChangeMeta()`, which was ignored when values were recoded into each other (#104)
 * fix unclear documentation in `removeValLabels()` (#111)
+* fix `changeMissings()` bug which dropped a single, existing value label and threw an error (#121)
 
 # eatGADS 1.1.1
 ## new features